Simple Rig: 2.0 WIP #163
